services: Add queue service
Add a Redis-backed queue service for IPC.

Eventually, users will be able to issue commands for specific
devices via the REST API. These commands will need to appear as
CoAP resources. As a multi-process architecture is being used,
i.e. one process for CoAP and one for REST, a queue primitive
is required for IPC. This commit adds such a primitive, using a
Redis server for data-sharing between processes.