services: Refactor RedisQueue into singleton
Refactor the RedisQueue implementation to improve better
separation of concerns and allow for a consistent service API.

This commit modifies the RedisQueue class to match the pattern
of the DatabaseSessionManager class. Notably, users of the queue
must now call an `init` function before initialisation.
Additionally, a singleton queue instance is defined in the
queue service.